experiments/us-modularity/pyproject.toml
[project]
name = "us-modularity"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "networkx>=3.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "scipy>=1.10.0",
    "duckdb>=1.0.0",
    "tqdm>=4.65.0",
    "pyyaml>=6.0",
]
